reggie-app-runner/pyproject.toml
[build-system]
requires = [
    "uv_build>=0.8.23,<0.9.0",
]
build-backend = "uv_build"

[project]
name = "reggie-app-runner"
version = "0.0.1+gae50b1b"
requires-python = ">=3.10"
dependencies = [
    "reggie-core @ file://${PROJECT_ROOT}/../reggie-core",
    "filelock",
    "dynaconf",
    "pyaml",
    "sh",
    "psutil",
    "requirements-parser",
    "giturlparse",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.uv.sources.reggie-core]
workspace = true
